remove dead spaces code in DAV layer
