+++ changelog/unreleased/remove-dead-code.md
@@ -1,0 +1,5 @@
+Enhancement: remove dead code
+
+Removed the unused SpacesHandler and related methods from the DAV layer
+
+https://github.com/cs3org/reva/pull/5279
